pyproject.toml
[tool.poetry]
name = "pytorch_advanced"
version = "0.1.0"
description = "For the pytorch deeplearning book"
authors = ["Shota Shimizu <shota_shimizu@caddi.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "*"
torchvision = "*" 
torchaudio = "*"
matplotlib = "*"
jupyterlab = "^3.0"
Pillow = "^8.4.0"
numpy = "*"
tqdm = "*"
python-dotenv="*"

[tool.poetry.dev-dependencies]
pytest = "*"
black = "^21.12b0"
blackcellmagic = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
